refactor: rename RNS class filename
